fix bug: the second wallet getblock, response Gson pare err

+++ app/src/main/java/chinapex/com/wallet/bean/response/ResponseGetBlock.java
@@ -11,14 +11,14 @@ public class ResponseGetBlock {
     /**
      * jsonrpc : 2.0
      * id : 1
-     * result : {"hash":"0x65e70b74a56122a8f6b98d1ac95116690d6e50fc644f511a6660d9140c8bcb85",
-     * "size":948,"version":0,
-     * "previousblockhash":"0x18f59297569eef5695612afa2a00cba1fe31de71810de49d065ed634f4c86a2c",
-     * "merkleroot":"0x0740e1926ab751f2081131d96e5f101d18ef1088a06cce52a35d4dd4c9927a7f",
-     * "time":1526537679,"index":1471562,"nonce":"ccead94b4055c66c",
+     * result : {"hash":"0x7a05b2f829ba56d35d502915cab72fb436c6ed3c7d59ccf940314ee6a86b7534",
+     * "size":888,"version":0,
+     * "previousblockhash":"0xed014f3d9f7660639eb09fbd93fcb78bd5d0fe1b83073cab310f1a6880d841da",
+     * "merkleroot":"0x60be173b7219e9c19557911b1c027b75ff82d9fe601eeeb1efcaa8f89a08950a",
+     * "time":1527043984,"index":1487935,"nonce":"2e970bd6cc3e5f4f",
      * "nextconsensus":"AaRDxTygdb5xYk1fku7uz7L5aTgNtn3J1w",
      * "script":{"invocation
-     * ":"40a5a8e653a8322072c37fd69a9e1a8a59b88112208c5a14535014d4bfb6508786d6601d2962bde809747859295a12c8dda8e93fa196c7a6e9871c48465b8ceb9740bf35665a231ba78c0989e740b71ff63d69a0c4329fe1c12469ce131f2fa8ff50457e86ac3454b83f957f666dd752ceb9c28eea149421eef8f0eaa0e11b305adb40bbe022f60a9673883b21aa8af486a9a7b3281b452cea01ceec1a68bb8a92836dccb578faa3922732a2a80489521da6760186342f4e577cb7e881264b2ffcf3a440ce42625054bdb0a25bb242df876afc49dab524f72faad1043bf6b3fe9e756cb8031de86897a8e78ebf8b0105f50f97b09375779d34a9a4451427993cb6aa81364064854e1f2d806bfd1734755570f120ea9c63d98a4a2ffa4332e7d10a8c119125fae142a0e8933936c71566545a6deeeea6ed358e1cfcbdadc8deb383d89107ae","verification":"55210209e7fd41dfb5c2f8dc72eb30358ac100ea8c72da18847befe06eade68cebfcb9210327da12b5c40200e9f65569476bbff2218da4f32548ff43b6387ec1416a231ee8210344925b6126c8ae58a078b5b2ce98de8fff15a22dac6f57ffd3b108c72a0670d121025bdf3f181f53e9696227843950deb72dcd374ded17c057159513c3d0abe20b6421026ce35b29147ad09e4afe4ec4a7319095f08198fa8babbe3c56e970b143528d2221039dafd8571a641058ccc832c5e2111ea39b09c0bde36050914384f7a48bce9bf92103c089d7122b840a4935234e82e26ae5efd0c2acb627239dc9f207311337b6f2c157ae"},"tx":[{"txid":"0x0d40a9166a0f5e7c3b90d1d05681944023a1718a2e3048bf2881e0e1ab5149e1","size":10,"type":"MinerTransaction","version":0,"attributes":[],"vin":[],"vout":[],"sys_fee":"0","net_fee":"0","scripts":[],"nonce":1079363180},{"txid":"0xdd2a1f010afd43535d2a40d3a321dc5b903af59a0ad04b7da090ee11cd33e549","size":262,"type":"ContractTransaction","version":0,"attributes":[],"vin":[{"txid":"0xa976fe69a563f429485d0732016b6c939cf2dab13aaac01f07ee20aa70847457","vout":1}],"vout":[{"n":0,"asset":"0xc56f33fc6ecfcd0c225c4ab356fee59390af8560be0e930faebe74a6daff7c9b","value":"6","address":"AaHQXZe1eQicV7bfjyDNQEjAtSHXFfMRP6"},{"n":1,"asset":"0xc56f33fc6ecfcd0c225c4ab356fee59390af8560be0e930faebe74a6daff7c9b","value":"378","address":"ARF9CNWdtpWZhMBvsbRd3jo12s4YcfeQyR"}],"sys_fee":"0","net_fee":"0","scripts":[{"invocation":"40666b5637ad044432e324382ee45ee709fdd6a0e303c762190b94b44fec13db89d678f86e81005c48a36cb7a50ec6e543085cf90aa56d7c967eec5483df6e7fb7","verification":"21024e56d7e0545da86eac54b7de54d398fbfc5def59e6c89dd4702ec38a4618ca13ac"}]}],"confirmations":14459,"nextblockhash":"0x261b1e31fc9c1e84300dc334a8a2cc7106a891ad97584202634999983e6231f2"}
+     * ":"403e719b298e357d581f7c67769428a1bc5b549b2b6437cb6faf6f96bd70e7721cec30d5bf9e7195817d946f7050e449e640c326023e24c6e38a943f5bf515291640bd5e89eb6c868c6ed43bebe20728e3314301b7eef0dea9d8274e8784bb82131347a80c586b092daf7788f9e65bc93b4e31192f149576c5894e3b30d8c46b02c940db599eda2718de39c6820f67d6f0607afdb80cde80ba42d205771d29972d2393a194ee7c0f130f8a094b90f143341d50dd6be6e66303230229ffe0fe27eaee2740524040bca3b90142a3a01da980b992a15e0c64e90bdb5c438650aba822515dc35fbca0d39baa2cc35b0e9eef01c046878e1706b7c5ca4e18f8f29d53825c62cb40957eb50e4cd845e49a552c7b8816e08c549616bb689e0f7cbd77a05cc7ada21444712ba3b172a85b24ce65993806e0cba06bfddf56c311eb2482a3376fde1046","verification":"55210209e7fd41dfb5c2f8dc72eb30358ac100ea8c72da18847befe06eade68cebfcb9210327da12b5c40200e9f65569476bbff2218da4f32548ff43b6387ec1416a231ee8210344925b6126c8ae58a078b5b2ce98de8fff15a22dac6f57ffd3b108c72a0670d121025bdf3f181f53e9696227843950deb72dcd374ded17c057159513c3d0abe20b6421026ce35b29147ad09e4afe4ec4a7319095f08198fa8babbe3c56e970b143528d2221039dafd8571a641058ccc832c5e2111ea39b09c0bde36050914384f7a48bce9bf92103c089d7122b840a4935234e82e26ae5efd0c2acb627239dc9f207311337b6f2c157ae"},"tx":[{"txid":"0xf3be55e43ff279fc904393884da19aabe13533df0afdc8b376e031fb0a1f1372","size":10,"type":"MinerTransaction","version":0,"attributes":[],"vin":[],"vout":[],"sys_fee":"0","net_fee":"0","scripts":[],"nonce":3426639695},{"txid":"0x24fe3fce8b65973ac3aba4a174e93f27d31618a3802f7c0faa1698a7bea24f96","size":202,"type":"ContractTransaction","version":0,"attributes":[],"vin":[{"txid":"0xfa257868244839897f7f607f9eab804c0ed93073f9eba4a66941efc35ea9b58b","vout":0}],"vout":[{"n":0,"asset":"0xc56f33fc6ecfcd0c225c4ab356fee59390af8560be0e930faebe74a6daff7c9b","value":"1","address":"AGh2PSiD5WVcrQJyywikGWmBG9r8CVrPB5"}],"sys_fee":"0","net_fee":"0","scripts":[{"invocation":"40dca5046ba538bee8cc4fe5ae343714f4f4b60b0a3f0f6298e4dc72b1d3e65f5a113ea5db7cca7f915656f703f78983358358b298b1609945a5492034295d8dff","verification":"210283c09741edf085b971317268ed2a782953651c8531782306986fa1e9ac4b5520ac"}]}],"confirmations":3,"nextblockhash":"0x724c4bcaf4ce353fc84bd0fb62723a7233cfdf5bd0108f0f67ec3e462b888ed7"}
      */
 
     private String jsonrpc;
@@ -51,33 +51,31 @@ public class ResponseGetBlock {
 
     public static class ResultBean {
         /**
-         * hash : 0x65e70b74a56122a8f6b98d1ac95116690d6e50fc644f511a6660d9140c8bcb85
-         * size : 948
+         * hash : 0x7a05b2f829ba56d35d502915cab72fb436c6ed3c7d59ccf940314ee6a86b7534
+         * size : 888
          * version : 0
-         * previousblockhash : 0x18f59297569eef5695612afa2a00cba1fe31de71810de49d065ed634f4c86a2c
-         * merkleroot : 0x0740e1926ab751f2081131d96e5f101d18ef1088a06cce52a35d4dd4c9927a7f
-         * time : 1526537679
-         * index : 1471562
-         * nonce : ccead94b4055c66c
+         * previousblockhash : 0xed014f3d9f7660639eb09fbd93fcb78bd5d0fe1b83073cab310f1a6880d841da
+         * merkleroot : 0x60be173b7219e9c19557911b1c027b75ff82d9fe601eeeb1efcaa8f89a08950a
+         * time : 1527043984
+         * index : 1487935
+         * nonce : 2e970bd6cc3e5f4f
          * nextconsensus : AaRDxTygdb5xYk1fku7uz7L5aTgNtn3J1w
          * script :
-         * {"invocation":"40a5a8e653a8322072c37fd69a9e1a8a59b88112208c5a14535014d4bfb6508786d6601d2962bde809747859295a12c8dda8e93fa196c7a6e9871c48465b8ceb9740bf35665a231ba78c0989e740b71ff63d69a0c4329fe1c12469ce131f2fa8ff50457e86ac3454b83f957f666dd752ceb9c28eea149421eef8f0eaa0e11b305adb40bbe022f60a9673883b21aa8af486a9a7b3281b452cea01ceec1a68bb8a92836dccb578faa3922732a2a80489521da6760186342f4e577cb7e881264b2ffcf3a440ce42625054bdb0a25bb242df876afc49dab524f72faad1043bf6b3fe9e756cb8031de86897a8e78ebf8b0105f50f97b09375779d34a9a4451427993cb6aa81364064854e1f2d806bfd1734755570f120ea9c63d98a4a2ffa4332e7d10a8c119125fae142a0e8933936c71566545a6deeeea6ed358e1cfcbdadc8deb383d89107ae","verification":"55210209e7fd41dfb5c2f8dc72eb30358ac100ea8c72da18847befe06eade68cebfcb9210327da12b5c40200e9f65569476bbff2218da4f32548ff43b6387ec1416a231ee8210344925b6126c8ae58a078b5b2ce98de8fff15a22dac6f57ffd3b108c72a0670d121025bdf3f181f53e9696227843950deb72dcd374ded17c057159513c3d0abe20b6421026ce35b29147ad09e4afe4ec4a7319095f08198fa8babbe3c56e970b143528d2221039dafd8571a641058ccc832c5e2111ea39b09c0bde36050914384f7a48bce9bf92103c089d7122b840a4935234e82e26ae5efd0c2acb627239dc9f207311337b6f2c157ae"}
-         * tx : [{"txid":"0x0d40a9166a0f5e7c3b90d1d05681944023a1718a2e3048bf2881e0e1ab5149e1",
+         * {"invocation":"403e719b298e357d581f7c67769428a1bc5b549b2b6437cb6faf6f96bd70e7721cec30d5bf9e7195817d946f7050e449e640c326023e24c6e38a943f5bf515291640bd5e89eb6c868c6ed43bebe20728e3314301b7eef0dea9d8274e8784bb82131347a80c586b092daf7788f9e65bc93b4e31192f149576c5894e3b30d8c46b02c940db599eda2718de39c6820f67d6f0607afdb80cde80ba42d205771d29972d2393a194ee7c0f130f8a094b90f143341d50dd6be6e66303230229ffe0fe27eaee2740524040bca3b90142a3a01da980b992a15e0c64e90bdb5c438650aba822515dc35fbca0d39baa2cc35b0e9eef01c046878e1706b7c5ca4e18f8f29d53825c62cb40957eb50e4cd845e49a552c7b8816e08c549616bb689e0f7cbd77a05cc7ada21444712ba3b172a85b24ce65993806e0cba06bfddf56c311eb2482a3376fde1046","verification":"55210209e7fd41dfb5c2f8dc72eb30358ac100ea8c72da18847befe06eade68cebfcb9210327da12b5c40200e9f65569476bbff2218da4f32548ff43b6387ec1416a231ee8210344925b6126c8ae58a078b5b2ce98de8fff15a22dac6f57ffd3b108c72a0670d121025bdf3f181f53e9696227843950deb72dcd374ded17c057159513c3d0abe20b6421026ce35b29147ad09e4afe4ec4a7319095f08198fa8babbe3c56e970b143528d2221039dafd8571a641058ccc832c5e2111ea39b09c0bde36050914384f7a48bce9bf92103c089d7122b840a4935234e82e26ae5efd0c2acb627239dc9f207311337b6f2c157ae"}
+         * tx : [{"txid":"0xf3be55e43ff279fc904393884da19aabe13533df0afdc8b376e031fb0a1f1372",
          * "size":10,"type":"MinerTransaction","version":0,"attributes":[],"vin":[],"vout":[],
-         * "sys_fee":"0","net_fee":"0","scripts":[],"nonce":1079363180},
-         * {"txid":"0xdd2a1f010afd43535d2a40d3a321dc5b903af59a0ad04b7da090ee11cd33e549",
-         * "size":262,"type":"ContractTransaction","version":0,"attributes":[],
-         * "vin":[{"txid":"0xa976fe69a563f429485d0732016b6c939cf2dab13aaac01f07ee20aa70847457",
-         * "vout":1}],"vout":[{"n":0,
+         * "sys_fee":"0","net_fee":"0","scripts":[],"nonce":3426639695},
+         * {"txid":"0x24fe3fce8b65973ac3aba4a174e93f27d31618a3802f7c0faa1698a7bea24f96",
+         * "size":202,"type":"ContractTransaction","version":0,"attributes":[],
+         * "vin":[{"txid":"0xfa257868244839897f7f607f9eab804c0ed93073f9eba4a66941efc35ea9b58b",
+         * "vout":0}],"vout":[{"n":0,
          * "asset":"0xc56f33fc6ecfcd0c225c4ab356fee59390af8560be0e930faebe74a6daff7c9b",
-         * "value":"6","address":"AaHQXZe1eQicV7bfjyDNQEjAtSHXFfMRP6"},{"n":1,
-         * "asset":"0xc56f33fc6ecfcd0c225c4ab356fee59390af8560be0e930faebe74a6daff7c9b",
-         * "value":"378","address":"ARF9CNWdtpWZhMBvsbRd3jo12s4YcfeQyR"}],"sys_fee":"0",
+         * "value":"1","address":"AGh2PSiD5WVcrQJyywikGWmBG9r8CVrPB5"}],"sys_fee":"0",
          * "net_fee":"0",
          * "scripts":[{"invocation
-         * ":"40666b5637ad044432e324382ee45ee709fdd6a0e303c762190b94b44fec13db89d678f86e81005c48a36cb7a50ec6e543085cf90aa56d7c967eec5483df6e7fb7","verification":"21024e56d7e0545da86eac54b7de54d398fbfc5def59e6c89dd4702ec38a4618ca13ac"}]}]
-         * confirmations : 14459
-         * nextblockhash : 0x261b1e31fc9c1e84300dc334a8a2cc7106a891ad97584202634999983e6231f2
+         * ":"40dca5046ba538bee8cc4fe5ae343714f4f4b60b0a3f0f6298e4dc72b1d3e65f5a113ea5db7cca7f915656f703f78983358358b298b1609945a5492034295d8dff","verification":"210283c09741edf085b971317268ed2a782953651c8531782306986fa1e9ac4b5520ac"}]}]
+         * confirmations : 3
+         * nextblockhash : 0x724c4bcaf4ce353fc84bd0fb62723a7233cfdf5bd0108f0f67ec3e462b888ed7
          */
 
         private String hash;
@@ -201,7 +199,7 @@ public class ResponseGetBlock {
         public static class ScriptBean {
             /**
              * invocation :
-             * 40a5a8e653a8322072c37fd69a9e1a8a59b88112208c5a14535014d4bfb6508786d6601d2962bde809747859295a12c8dda8e93fa196c7a6e9871c48465b8ceb9740bf35665a231ba78c0989e740b71ff63d69a0c4329fe1c12469ce131f2fa8ff50457e86ac3454b83f957f666dd752ceb9c28eea149421eef8f0eaa0e11b305adb40bbe022f60a9673883b21aa8af486a9a7b3281b452cea01ceec1a68bb8a92836dccb578faa3922732a2a80489521da6760186342f4e577cb7e881264b2ffcf3a440ce42625054bdb0a25bb242df876afc49dab524f72faad1043bf6b3fe9e756cb8031de86897a8e78ebf8b0105f50f97b09375779d34a9a4451427993cb6aa81364064854e1f2d806bfd1734755570f120ea9c63d98a4a2ffa4332e7d10a8c119125fae142a0e8933936c71566545a6deeeea6ed358e1cfcbdadc8deb383d89107ae
+             * 403e719b298e357d581f7c67769428a1bc5b549b2b6437cb6faf6f96bd70e7721cec30d5bf9e7195817d946f7050e449e640c326023e24c6e38a943f5bf515291640bd5e89eb6c868c6ed43bebe20728e3314301b7eef0dea9d8274e8784bb82131347a80c586b092daf7788f9e65bc93b4e31192f149576c5894e3b30d8c46b02c940db599eda2718de39c6820f67d6f0607afdb80cde80ba42d205771d29972d2393a194ee7c0f130f8a094b90f143341d50dd6be6e66303230229ffe0fe27eaee2740524040bca3b90142a3a01da980b992a15e0c64e90bdb5c438650aba822515dc35fbca0d39baa2cc35b0e9eef01c046878e1706b7c5ca4e18f8f29d53825c62cb40957eb50e4cd845e49a552c7b8816e08c549616bb689e0f7cbd77a05cc7ada21444712ba3b172a85b24ce65993806e0cba06bfddf56c311eb2482a3376fde1046
              * verification :
              * 55210209e7fd41dfb5c2f8dc72eb30358ac100ea8c72da18847befe06eade68cebfcb9210327da12b5c40200e9f65569476bbff2218da4f32548ff43b6387ec1416a231ee8210344925b6126c8ae58a078b5b2ce98de8fff15a22dac6f57ffd3b108c72a0670d121025bdf3f181f53e9696227843950deb72dcd374ded17c057159513c3d0abe20b6421026ce35b29147ad09e4afe4ec4a7319095f08198fa8babbe3c56e970b143528d2221039dafd8571a641058ccc832c5e2111ea39b09c0bde36050914384f7a48bce9bf92103c089d7122b840a4935234e82e26ae5efd0c2acb627239dc9f207311337b6f2c157ae
              */
@@ -228,7 +226,7 @@ public class ResponseGetBlock {
 
         public static class TxBean {
             /**
-             * txid : 0x0d40a9166a0f5e7c3b90d1d05681944023a1718a2e3048bf2881e0e1ab5149e1
+             * txid : 0xf3be55e43ff279fc904393884da19aabe13533df0afdc8b376e031fb0a1f1372
              * size : 10
              * type : MinerTransaction
              * version : 0
@@ -238,7 +236,7 @@ public class ResponseGetBlock {
              * sys_fee : 0
              * net_fee : 0
              * scripts : []
-             * nonce : 1079363180
+             * nonce : 3426639695
              */
 
             private String txid;
@@ -247,7 +245,7 @@ public class ResponseGetBlock {
             private int version;
             private String sys_fee;
             private String net_fee;
-            private int nonce;
+            private long nonce;
             private List<?> attributes;
             private List<?> vin;
             private List<?> vout;
@@ -301,11 +299,11 @@ public class ResponseGetBlock {
                 this.net_fee = net_fee;
             }
 
-            public int getNonce() {
+            public long getNonce() {
                 return nonce;
             }
 
-            public void setNonce(int nonce) {
+            public void setNonce(long nonce) {
                 this.nonce = nonce;
             }
